# txtclaw-web/public/openapi.yaml
openapi: 3.1.0
info:
  title: TXT CLAW API
  version: v1
  description: |
    TXT CLAW developer API (preview).

    Create a dedicated OpenClaw agent and talk to it over HTTPS.
    SMS provisioning is a separate/optional lane.
servers:
  - url: https://txtclaw-sms-e2e.lopez731.workers.dev
    description: Preview (Workers)
security:
  - bearerAuth: []
paths:
  /v1/status:
    get:
      summary: Status
      operationId: getStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Payment required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: TXTCLAW_API_KEY
  schemas:
    ErrorResponse:
      type: object
      description: Error response. Always include `trace_id` when reporting issues.
      additionalProperties: true
      properties:
        error: { type: string }
        trace_id: { type: string }
        limit_key: { type: string }
        reason: { type: string }
      required: [error]
    StatusResponse:
      type: object
      additionalProperties: false
      properties:
        ok: { type: boolean }
        service: { type: string }
        version: { type: string }
        mode: { type: string }
        trace_id: { type: string }
      required: [ok, service, version, mode, trace_id]
...